update T33250: 1. Change the CRLF to LF 2. Update case name to T33250_idic_KCSSensorList.py
